+++ apolloschurchapp/android/app/BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.apolloschurchapp",
+    package = "com.differential.apollos",
 )
 
 android_resource(
     name = "res",
-    package = "com.apolloschurchapp",
+    package = "com.differential.apollos",
     res = "src/main/res",
 )
