+++ apolloschurchapp/android/app/build.gradle
@@ -151,7 +151,7 @@ android {
         ndk {
             abiFilters "armeabi-v7a", "x86", "arm64-v8a", "x86_64"
         }
-        resValue "string", "build_config_package", "com.apolloschurchapp"
+        resValue "string", "build_config_package", "com.differential.apollos"
         multiDexEnabled true
     }
     signingConfigs {
