+++ apolloschurchapp/android/app/src/main/java/com/differential/apollos/MainApplication.java
@@ -1,4 +1,4 @@
-package com.apolloschurchapp;
+package com.differential.apollos;
 
 import android.app.Application;
 import android.util.Log;
